pyproject.toml
[project]
name = "ai-mentalist"
version = "0.1.0"
description = "A virtual mentalist chatbot using RAG over scanned mentalism PDF books"
requires-python = ">=3.11"
dependencies = [
    "langchain>=0.3",
    "langchain-community>=0.3",
    "langchain-openai>=0.3",
    "openai>=1.0",
    "faiss-cpu>=1.9,<1.10",
    "pytesseract>=0.3",
    "pdf2image>=1.17",
    "streamlit>=1.40",
    "python-dotenv>=1.0",
]
